test: 인수테스트 중복 제거
